[Refactor] standardization - move all the extension-related files under `./src/editor/contrib/`
