+++ src/editor/contrib/commandExtension/commandExtension.ts
@@ -3,10 +3,10 @@ import { EditorView } from "prosemirror-view";
 import { trySafe } from "src/base/common/error";
 import { Shortcut } from "src/base/common/keyboard";
 import { ILogService } from "src/base/common/logger";
-import { EditorExtensionIDs } from "src/editor/common/extension/builtInExtension";
+import { EditorExtensionIDs } from "src/editor/contrib/builtInExtensionList";
 import { EditorExtension, IEditorExtension } from "src/editor/common/extension/editorExtension";
 import { IEditorWidget } from "src/editor/editorWidget";
-import { registerBasicEditorCommands } from "src/editor/view/contrib/editorCommands";
+import { registerBasicEditorCommands } from "src/editor/contrib/commandExtension/editorCommands";
 import { Command } from "src/platform/command/common/command";
 import { ICommandService } from "src/platform/command/common/commandService";
 import { RegistrantType } from "src/platform/registrant/common/registrant";
